refactor : camera init
