+++ Assets/Scripts/Game/Family/FamilyScene.cs
@@ -17,12 +17,14 @@ namespace Game.FM
             GameObject playerObject = Instantiate(mainCharacterPrefab, Vector3.up * 5, Quaternion.identity);
             // playerObject.transform.SetParent(GameObject.Find("ObjectRoot").transform);
             PlayerManager.Instance.player = playerObject.GetComponent<Player>();
-        
-            CameraManager.Instance.virtualCamera = GameObject.Find("followCam").GetComponent<CinemachineVirtualCamera>();
+
+            GameObject virtualCam = GameObject.Find("followCam");
+            CameraManager.Instance.fermiCamera = virtualCam.GetComponent<FermiCamera>();
+            CameraManager.Instance.virtualCamera = virtualCam.GetComponent<CinemachineVirtualCamera>();
             CameraManager.Instance.cinemachineOrbitalTransposer = CameraManager.Instance.virtualCamera.GetCinemachineComponent<CinemachineOrbitalTransposer>();
             CameraManager.Instance.virtualCamera.LookAt = playerObject.transform;
             CameraManager.Instance.virtualCamera.Follow = playerObject.transform;
-        
+
             GameObject touchPadPrefab = UIManager.Instance.ShowSceneUI("Prefabs/UI/TouchPadCanvas");
             GameManager.Instance.GameScene = this;
         }
